backend/requirements.txt
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
google-generativeai==0.3.2
scikit-learn==1.4.1
numpy==1.26.4
pandas==2.2.1
python-jose==3.3.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
Pillow==10.2.0 